Integrate resource resolution cases into existing schema tests
