+++ tests/main/jsonschema/test_main_jsonschema.py
@@ -17,13 +17,17 @@ import warnings
 from collections import UserDict, defaultdict
 from collections.abc import Callable as ABCCallable
 from collections.abc import Sequence
-from contextlib import nullcontext
 from dataclasses import Field as DataclassField
 from decimal import Decimal
 from operator import itemgetter
 from pathlib import Path, PurePath
 from types import MappingProxyType
+from contextlib import ExitStack, nullcontext
+from functools import partial
+from http.server import ThreadingHTTPServer
+from threading import Thread
 from typing import TYPE_CHECKING, get_args, get_type_hints
+from urllib.parse import urlparse
 
 import black
 import msgspec
@@ -34,6 +38,9 @@ from jsonschema import ValidationError as SchemaValidationError
 from packaging import version
 from pydantic import VERSION as PYDANTIC_VERSION
 from pydantic import PydanticUserError, ValidationError
+from jsonschema.validators import validator_for
+from referencing import Registry, Resource
+from referencing.exceptions import Unresolvable
 
 from datamodel_code_generator import (
     MIN_VERSION,
@@ -41,6 +48,7 @@ from datamodel_code_generator import (
     DataModelType,
     DefaultValueType,
     Error,
+    GenerateConfig,
     InputFileType,
     InvalidFileFormatError,
     PythonVersion,
@@ -71,6 +79,7 @@ from tests.conftest import (
     MockHttpxResponse,
     assert_directory_content,
     assert_httpx_get_kwargs,
+    assert_inputs_not_mutated,
     assert_mutable_copy_is_isolated,
     assert_no_uncommented_generated_code,
     assert_output,
@@ -80,6 +89,7 @@ from tests.conftest import (
     freeze_time,
     validate_generated_code,
 )
+from tests.data.python.schema_resource_server import RecordingRequestHandler
 from tests.main.conftest import (
     ALIASES_DATA_PATH,
     BACKEND_GOLDEN_CASES,
@@ -22048,3 +22058,236 @@ def test_schema_validator_legacy_custom_helper(
         for container in containers:
             model.model_validate(container(values["valid"]))
             _assert_model_json_invalid(model.model_validate, container(values["invalid"]), "value_error")
+
+
+RESOURCES_EXPECTED = EXPECTED_JSON_SCHEMA_PATH / "nested_resources"
+
+
+RESOURCES_CASES = json.loads((RESOURCES_EXPECTED / "cases.json").read_text(encoding="utf-8"))
+
+
+@pytest.mark.parametrize(
+    ("case", "entrypoint"),
+    [
+        *(
+            (case, entrypoint)
+            for case in RESOURCES_CASES
+            for entrypoint in ("cli", "api", "dict", "custom-cli", "custom-api")
+        ),
+        *(
+            (case, entrypoint)
+            for case in RESOURCES_CASES
+            if "http_requests" in RESOURCES_CASES[case]
+            for entrypoint in ("url-cli", "url-api")
+        ),
+    ],
+)
+@pytest.mark.parametrize("formatter", ["external", "builtin"])
+def test_nested_schema_resources(tmp_path: Path, case: str, entrypoint: str, formatter: str) -> None:
+    """Check native semantics, actual fetches, generated bytes and unchanged caller input."""
+    source = JSON_SCHEMA_DATA_PATH / "nested_resources" / case / "root.json"
+    schema = json.loads(source.read_text(encoding="utf-8"))
+    validator_class = validator_for(schema)
+    validator_class.check_schema(schema)
+    registry = Registry().with_resources(
+        (relative, Resource.from_contents(json.loads((source.parent / relative).read_text(encoding="utf-8"))))
+        for relative in RESOURCES_CASES[case]["resources"]
+    )
+    validator = validator_class(schema, registry=registry)
+    output = tmp_path / "output.py"
+    formatters = ["builtin"] if formatter == "builtin" else ["black", "isort"]
+    requests = []
+    with ExitStack() as stack:
+        input_value = source
+        url = ""
+        if entrypoint == "dict":
+            stack.enter_context(chdir(source.parent))
+        if entrypoint.startswith("url-"):
+            server = stack.enter_context(
+                ThreadingHTTPServer(
+                    ("127.0.0.1", 0),
+                    partial(RecordingRequestHandler, requests=requests, directory=str(source.parent)),
+                )
+            )
+            thread = Thread(target=server.serve_forever, kwargs={"poll_interval": 0.01})
+            thread.start()
+            stack.callback(thread.join)
+            stack.callback(server.shutdown)
+            url = f"http://127.0.0.1:{server.server_port}/root.json"
+            input_value = urlparse(url)
+        expected_error = RESOURCES_CASES[case].get("generation_errors", {}).get(entrypoint)
+        with (
+            assert_inputs_not_mutated({"schema": schema}),
+            warnings.catch_warnings(record=True) as recorded_warnings,
+            pytest.raises(Error) if expected_error else nullcontext() as exception,
+        ):
+            warnings.simplefilter("always", UserWarning)
+            if entrypoint.endswith("cli"):
+                run_main_with_args([
+                    *(["--url", url] if entrypoint == "url-cli" else ["--input", str(source)]),
+                    "--output",
+                    str(output),
+                    "--input-file-type",
+                    "jsonschema",
+                    "--disable-timestamp",
+                    "--custom-file-header-path",
+                    str(DATA_PATH / "custom_file_header.txt"),
+                    "--allow-remote-refs" if entrypoint.startswith("url-") else "--no-allow-remote-refs",
+                    "--allow-private-network",
+                    *(
+                        ["--custom-template-dir", str(DATA_PATH / "templates_nested_resources")]
+                        if entrypoint.startswith("custom-")
+                        else []
+                    ),
+                    "--formatters",
+                    *formatters,
+                ])
+            else:
+                generate(
+                    schema if entrypoint == "dict" else input_value,
+                    config=GenerateConfig(
+                        input_file_type=InputFileType.JsonSchema,
+                        output=output,
+                        input_filename="root.json",
+                        disable_timestamp=True,
+                        custom_file_header_path=DATA_PATH / "custom_file_header.txt",
+                        allow_remote_refs=entrypoint.startswith("url-"),
+                        allow_private_network=True,
+                        custom_template_dir=DATA_PATH / "templates_nested_resources"
+                        if entrypoint.startswith("custom-")
+                        else None,
+                        formatters=[Formatter(value) for value in formatters],
+                    ),
+                )
+        assert_output(
+            json.dumps([str(warning.message) for warning in recorded_warnings], indent=2),
+            RESOURCES_EXPECTED
+            / (
+                "oneof_minimum_warnings.txt"
+                if case.startswith("oneof") and PYDANTIC_VERSION in {"2.0", "2.0.0"}
+                else "no_warnings.txt"
+            ),
+        )
+        if exception is not None:
+            assert_output(
+                json.dumps(
+                    {
+                        "native": [validator.is_valid(payload) for payload in RESOURCES_CASES[case]["payloads"]],
+                        "generation_error": str(exception.value),
+                    },
+                    indent=2,
+                ),
+                RESOURCES_EXPECTED / expected_error,
+            )
+            return
+        assert_output(
+            output.read_text(encoding="utf-8"),
+            (RESOURCES_EXPECTED / "custom" if entrypoint.startswith("custom-") else RESOURCES_EXPECTED) / f"{case}.py",
+        )
+        if entrypoint.startswith("url-"):
+            assert_output(json.dumps(requests, indent=2), RESOURCES_EXPECTED / f"{case}_requests.txt")
+    records = {"native": [], "generated": []}
+    with _generated_model(output, "generated_nested_resource", "Root") as model:
+        for payload in RESOURCES_CASES[case]["payloads"]:
+            records["native"].append(validator.is_valid(payload))
+            with assert_inputs_not_mutated({"payload": payload}):
+                try:
+                    value = model.model_validate(payload)
+                except ValidationError:
+                    records["generated"].append("rejected")
+                else:
+                    records["generated"].append(value.model_dump(mode="json", by_alias=True))
+    assert_output(json.dumps(records, indent=2), RESOURCES_EXPECTED / f"{case}_runtime.txt")
+
+
+@pytest.mark.parametrize("entrypoint", ["cli", "api", "dict", "url-cli", "url-api"])
+@pytest.mark.parametrize("formatter", ["external", "builtin"])
+def test_missing_embedded_anchor(
+    tmp_path: Path, capsys: pytest.CaptureFixture[str], entrypoint: str, formatter: str
+) -> None:
+    """Reject a missing resource-local anchor without reading a physical decoy."""
+    source = JSON_SCHEMA_DATA_PATH / "nested_resources" / "missing_anchor" / "root.json"
+    schema = json.loads(source.read_text(encoding="utf-8"))
+    validator_class = validator_for(schema)
+    validator_class.check_schema(schema)
+    validator = validator_class(schema, registry=Registry().with_resource("root.json", Resource.from_contents(schema)))
+    payload = json.loads((RESOURCES_EXPECTED / "missing_anchor_payload.json").read_text(encoding="utf-8"))
+    with pytest.raises(Unresolvable) as native_error:
+        validator.is_valid(payload)
+    assert_output(str(native_error.value), RESOURCES_EXPECTED / "missing_anchor_native_error.txt")
+    output = tmp_path / "output.py"
+    formatters = ["builtin"] if formatter == "builtin" else ["black", "isort"]
+    requests = []
+    with ExitStack() as stack:
+        input_value = source
+        url = ""
+        if entrypoint == "dict":
+            stack.enter_context(chdir(source.parent))
+        if entrypoint.startswith("url-"):
+            server = stack.enter_context(
+                ThreadingHTTPServer(
+                    ("127.0.0.1", 0),
+                    partial(RecordingRequestHandler, requests=requests, directory=str(source.parent)),
+                )
+            )
+            thread = Thread(target=server.serve_forever, kwargs={"poll_interval": 0.01})
+            thread.start()
+            stack.callback(thread.join)
+            stack.callback(server.shutdown)
+            url = f"http://127.0.0.1:{server.server_port}/root.json"
+            input_value = urlparse(url)
+        with assert_inputs_not_mutated({"schema": schema}):
+            if entrypoint.endswith("cli"):
+                run_main_with_args(
+                    [
+                        *(["--url", url] if entrypoint == "url-cli" else ["--input", str(source)]),
+                        "--output",
+                        str(output),
+                        "--input-file-type",
+                        "jsonschema",
+                        "--allow-remote-refs" if entrypoint.startswith("url-") else "--no-allow-remote-refs",
+                        "--allow-private-network",
+                        "--formatters",
+                        *formatters,
+                    ],
+                    expected_exit=Exit.ERROR,
+                    capsys=capsys,
+                    expected_stderr=(RESOURCES_EXPECTED / "missing_anchor_cli_error.txt").read_text(encoding="utf-8"),
+                )
+            else:
+                with pytest.raises(Error) as error:
+                    generate(
+                        schema if entrypoint == "dict" else input_value,
+                        config=GenerateConfig(
+                            input_file_type=InputFileType.JsonSchema,
+                            output=output,
+                            allow_remote_refs=entrypoint.startswith("url-"),
+                            allow_private_network=True,
+                            formatters=[Formatter(value) for value in formatters],
+                        ),
+                    )
+                assert_output(str(error.value), RESOURCES_EXPECTED / "missing_anchor_api_error.txt")
+        if entrypoint.startswith("url-"):
+            assert_output(json.dumps(requests, indent=2), RESOURCES_EXPECTED / "missing_anchor_requests.txt")
+
+
+@pytest.mark.parametrize("formatter", ["external", "builtin"])
+def test_repeated_schema_documents(tmp_path: Path, formatter: str) -> None:
+    """Keep the active resource owner when a source list revisits a physical document."""
+    source = JSON_SCHEMA_DATA_PATH / "nested_resources" / "duplicate_documents"
+    with chdir(source):
+        generated = JsonSchemaParser(
+            [source / "first.json", source / "second.json", source / "first.json"],
+            formatters=[Formatter.BUILTIN] if formatter == "builtin" else [Formatter.BLACK, Formatter.ISORT],
+        ).parse()
+    assert_output(json.dumps([list(path) for path in generated]), RESOURCES_EXPECTED / "duplicate_document_paths.txt")
+    for name in ("first", "second"):
+        result = generated[f"{name}.py",]
+        assert_output(result.body, RESOURCES_EXPECTED / "duplicate_documents" / f"{name}.py")
+        output = tmp_path / f"{name}.py"
+        output.write_text(result.body, encoding="utf-8")
+        with _generated_model(output, f"generated_{name}", name.title()) as model:
+            assert_output(
+                json.dumps(model.model_validate({"value": "ok"}).model_dump(mode="json")),
+                RESOURCES_EXPECTED / "duplicate_document_runtime.txt",
+            )
